requirements.txt
asgiref==3.2.7
astroid==2.3.3
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.4
django-oauth-toolkit==1.3.2
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
phonenumbers==8.12.1
Pillow==7.1.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
wrapt==1.11.2

# documentation dependencies
coreapi==2.3.3
coreschema==0.0.4
drf-yasg==1.17.1
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.3
pyparsing==2.4.7
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
uritemplate==3.0.1
